impl the order/limit/after in query_cells.
